% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{GetTableMetadata}
\alias{GetTableMetadata}
\title{FUNCTION_TITLE}
\usage{
GetTableMetadata(variable = fieldsAll, label = label_names)
}
\arguments{
\item{variable}{PARAM_DESCRIPTION, Default: fieldsAll}

\item{label}{PARAM_DESCRIPTION, Default: label_names}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
