% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{ReadData}
\alias{ReadData}
\title{FUNCTION_TITLE}
\usage{
ReadData()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
