% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_age.R
\name{tar_age}
\alias{tar_age}
\title{Calculate person age}
\usage{
tar_age(birth_date, x_date)
}
\arguments{
\item{birth_date}{A date value for example "1974-11-23"}

\item{x_date}{A date value until the age lasts for example "2023-02-16"}
}
\value{
A numeric value in (years)
}
\description{
This function calculates the age of a person by given birthdate and the end date
}
\examples{
tar_age("1974-11-23", "2023-02-16")

}
