% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bmi.R
\name{tar_bmi}
\alias{tar_bmi}
\title{Calculating BMI (kg/m^2)}
\usage{
tar_bmi(weight, height)
}
\arguments{
\item{weight}{Numeric, kg}

\item{height}{Numeric, cm}
}
\value{
A numeric vector containing the body mass index. If weight or
  height is missing, the corresponding result is NA.
}
\description{
Calculating BMI (kg/m^2)
}
\examples{
tar_bmi(weight = 100, height = 190)
tar_bmi(weight = c(80, NA), height = c(180, 175))
}
