% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_cast_data}
\alias{tar_cast_data}
\title{Cast registry data}
\usage{
tar_cast_data(data)
}
\arguments{
\item{data}{A list or data frame with registry fields.}
}
\value{
A one-row data frame with standard column types.
}
\description{
Coerce raw input into a one-row data frame with registry fields.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
