% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_color_duplicates.R
\name{tar_color_duplicates}
\alias{tar_color_duplicates}
\title{Color duplicate cells within specific column.}
\usage{
tar_color_duplicates(df, var)
}
\arguments{
\item{df}{The input data frame.}

\item{var}{The column to be used for coloring duplicates.}
}
\value{
A gt table with colored duplicate groups.
}
\description{
To visually identify this function will color duplicates in a column.
}
\examples{
# Example usage:
tar_color_duplicates(mtcars, cyl)

}
\author{
tarjae
}
