% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sumna.R
\name{tar_count_na}
\alias{tar_count_na}
\title{NA counter}
\usage{
tar_count_na(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
returns
}
\description{
Counts NA's across columns
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 tar_count_na(mtcars)
 }
}
}
