% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_create_data}
\alias{tar_create_data}
\title{Create registry record}
\usage{
tar_create_data(data)
}
\arguments{
\item{data}{A named list or data frame with registry fields.}
}
\value{
Invisible NULL.
}
\description{
Add a new record to the in-memory `responses` data frame.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
