% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_create_default_record}
\alias{tar_create_default_record}
\title{Create default registry record}
\usage{
tar_create_default_record()
}
\value{
A one-row data frame with default values.
}
\description{
Return an empty default registry record.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
