% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_delete_data}
\alias{tar_delete_data}
\title{Delete registry record}
\usage{
tar_delete_data(data)
}
\arguments{
\item{data}{A list or data frame containing an `id` field.}
}
\value{
Invisible NULL.
}
\description{
Delete a record from `responses` by id.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
