% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_flag.R
\name{tar_flag}
\alias{tar_flag}
\title{Create a combined flag plot}
\usage{
tar_flag(countries, name)
}
\arguments{
\item{countries}{A character vector of country codes (ISO 3166\-1 alpha\-2).}

\item{name}{A character string for the name of the study. This will be used
to construct the output file name.}
}
\value{
None. The function saves the combined flag image to a file.
}
\description{
This function downloads flag images of specified countries, combines them
vertically with spaces in between, and saves the combined image to a file.
}
\examples{
tar_flag(c("nl", "se", "es", "si", "dk", "no", "us"), "rapido")

}
