% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_get_next_id}
\alias{tar_get_next_id}
\title{Get next registry ID}
\usage{
tar_get_next_id()
}
\value{
Integer ID.
}
\description{
Get the next numeric ID based on the `responses` data frame.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
