% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_get_table_metadata}
\alias{tar_get_table_metadata}
\title{Registry field metadata}
\usage{
tar_get_table_metadata(variable = fieldsAll, label = label_names)
}
\arguments{
\item{variable}{Character vector of field names. Default: fieldsAll.}

\item{label}{Character vector of labels. Default: label_names.}
}
\value{
A list with element `fields`, a named character vector.
}
\description{
Create a named vector of fields and labels for the registry.
}
\details{
Expects `variable` and `label` to be the same length.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
