% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_get_value.R
\name{tar_get_value}
\alias{tar_get_value}
\title{Look-up function}
\usage{
<<<<<<< HEAD
tar_get_value(mykeyvalue)
}
\arguments{
\item{mykeyvalue}{an unquoted key or value}
=======
tar_get_value(mykey)
}
\arguments{
\item{mykey}{an unquoted key or value}
>>>>>>> 90c08958b66a7344fa5f094c3eb3a64690cbc130
}
\value{
a key or value, depending on what is wanted
}
\description{
Look-up function
}
\examples{
# Example data
<<<<<<< HEAD
=======
lookup_vector <- c(Hans = "1006341612", Maria = "1040405318")
>>>>>>> 90c08958b66a7344fa5f094c3eb3a64690cbc130
df <- data.frame(Name = c("Hans", "Maria"), azl = c("1006341612", "1040405318"), stringsAsFactors = FALSE)

tar_get_value(Hans)       # Returns "1006341612"
tar_get_value(1006341612) # Returns "Hans"
}
