% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_headtail.R
\name{tar_head_tail}
\alias{tar_head_tail}
\title{Show head and tail of data frame}
\usage{
tar_head_tail(data, nh = 5, nt = 5)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{nh}{PARAM_DESCRIPTION, Default: 5}

\item{nt}{PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{rownames}}
}
