% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_impute_outlier.R
\name{tar_impute_outlier}
\alias{tar_impute_outlier}
\title{Imputates outliers and shows 4 different method plots}
\usage{
tar_impute_outlier(df, variable)
}
\arguments{
\item{df}{data frame or tibble where the numeric variable lives}

\item{variable}{Numeric variable column in the data frame}
}
\value{
4 plots with mean, median, mode and capping imputation
}
\description{
Imputates outliers and shows 4 different method plots
}
