% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_oneline.R
\name{tar_oneline}
\alias{tar_oneline}
\title{tar_oneline}
\usage{
tar_oneline()
}
\value{
A character vector with the reformatted one-line code.
}
\description{
Creates a one-liner from a multiline code example in the clipboard.
}
\examples{
\dontrun{
if (interactive()) {
  # Your multiline string
  # copy this code
  numericInput(
    "Register_Nr",
    "Nr.",
    min = 0,
    max = 20,
    step = 1,
    value = 0
  )
}

# then apply
tar_oneline()
}
}
\seealso{
\code{\link[formatR]{tidy_source}}
}
