% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_read_data}
\alias{tar_read_data}
\title{Read registry data}
\usage{
tar_read_data()
}
\value{
A data frame or NULL if not available.
}
\description{
Return the in-memory `responses` data frame, if present.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
