% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_update_data}
\alias{tar_update_data}
\title{Update registry record}
\usage{
tar_update_data(data)
}
\arguments{
\item{data}{A named list or data frame with registry fields.}
}
\value{
Invisible NULL.
}
\description{
Update an existing record in `responses` by row name.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
