% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry.R
\name{tar_update_inputs}
\alias{tar_update_inputs}
\title{Update Shiny inputs from record}
\usage{
tar_update_inputs(data, session)
}
\arguments{
\item{data}{A one-row data frame with registry fields.}

\item{session}{Shiny session.}
}
\value{
Invisible NULL.
}
\description{
Push a registry record into Shiny input controls.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shiny]{updateTextInput}}, \code{\link[shiny]{updateNumericInput}}, \code{\link[shiny]{updateDateInput}}
}
