% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry_ui.R
\name{tar_WellPanel}
\alias{tar_WellPanel}
\title{tar_WellPanel}
\usage{
tar_WellPanel(n)
}
\arguments{
\item{n}{integer the gives the number of months in follow-up}
}
\value{
returns a wellpanel a quasi shinydashboard box
(without ShinyDashboard)
}
\description{
Creates a shinydashboard box equivalent with shiny
}
\details{
Builds a wellPanel containing follow-up inputs for the given month.
}
\examples{
\donttest{
if(interactive()){
 #EXAMPLE1
 }
}
}
