% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_animbar.R
\name{tar_animbar}
\alias{tar_animbar}
\title{Create Animated Bar Chart}
\usage{
tar_animbar(
  x,
  filename = "my_animated_bar",
  color1 = "steelblue1",
  color2 = "purple3",
  title = "\%",
  horizontal = FALSE
)
}
\arguments{
\item{x}{An integer indicating the maximum value for the bar chart.}

\item{filename}{A string specifying the name of the output file (default: "my_animated_bar").}

\item{color1}{A string specifying the color for the first part of the bar (default: "steelblue1").}

\item{color2}{A string specifying the color for the second part of the bar (default: "purple3").}

\item{title}{A string specifying the title suffix for the animation (default: "\%").}

\item{horizontal}{A logical value indicating whether the bar chart should be horizontal (default: FALSE).}
}
\value{
No return value. The function generates and saves an animated bar chart as a GIF file.
}
\description{
This function creates an animated bar chart with customizable colors and orientation. The resulting GIF can be saved to a specified filename.
}
\details{
This function generates an animated bar chart with customizable colors and orientation.
The resulting GIF can be saved to a specified filename.
}
\examples{
tar_animbar(x = 100, filename = "a3", color1 = "gold", color2 = "purple3", title = "\%", horizontal = TRUE)
tar_animbar(x = 45, filename = "a4", color1 = "steelblue", color2 = "purple3", title = "\%", horizontal = FALSE)
tar_animbar(x = 17, filename = "rapido_LRR", horizontal = FALSE)

}
