% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bmi.R
\name{tar_bmi}
\alias{tar_bmi}
\title{Calculating BMI (kg/m²)}
\usage{
tar_bmi(weight, height)
}
\arguments{
\item{weight}{Numeric, kg}

\item{height}{Numeric, meter}
}
\value{
body mass index BMI
}
\description{
Calculating BMI (kg/m²)
}
\examples{
weight = 100
height = 1.90
tar_bmi(weight, height)
}
