% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_bsa.R
\name{tar_bsa}
\alias{tar_bsa}
\title{Body surface area calculator (Dubois formula)}
\usage{
tar_bsa(height, weight)
}
\arguments{
\item{height}{A numeric value in cm}

\item{weight}{A numeric value in kg}
}
\value{
A numeric value in (m²)
}
\description{
Body surface area calculator (Dubois formula)
}
\examples{
tar_bsa(180, 80)

}
