% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_color_duplicates.R
\name{tar_color_duplicates}
\alias{tar_color_duplicates}
\title{Color Duplicate Cells in Specified Columns}
\usage{
tar_color_duplicates(df, vars)
}
\arguments{
\item{df}{A data frame in which to look for duplicates.}

\item{vars}{A vector of column names in which to look for duplicates.}
}
\value{
A `gt` table with colored duplicate groups.
}
\description{
This function colors cells in specified columns (n=x) of a dataframe that are duplicates.
It is useful for visually identifying duplicates in a `gt` table format.
}
\examples{
tar_color_duplicates(mtcars, c("cyl", "mpg"))

}
