% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sumna.R
\name{tar_count_na}
\alias{tar_count_na}
\title{NA counter}
\usage{
tar_count_na(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A named numeric vector with NA counts per column.
}
\description{
Counts NA values across columns.
}
\examples{
\dontrun{
if(interactive()){
 tar_count_na(mtcars)
 }
}
}
