% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_sumna.R
\name{tar_count_na}
\alias{tar_count_na}
\title{FUNCTION_TITLE}
\usage{
tar_count_na(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
