% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_hipec_registry_ui.R
\name{tar_dateInput}
\alias{tar_dateInput}
\title{tar_dateinput}
\usage{
tar_dateInput(inputId, label)
}
\arguments{
\item{inputId}{character}

\item{label}{character}
}
\value{
returns a date in dd-MM-yyyy format
}
\description{
Handles date format in Austrian HIPEC Registry(c)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[shinyWidgets]{airDatepicker}}
}
