% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_install.R
\name{tar_install}
\alias{tar_install}
\title{Short Installer}
\usage{
tar_install(x)
}
\arguments{
\item{x}{The name of the package you want to install.}
}
\value{
A character string with the corresponding `install.packages()` command.
}
\description{
Helper that returns an `install.packages()` command for a package name.
}
\details{
This function uses non-standard evaluation to allow you to specify
the package name unquoted.
}
\examples{
\donttest{
tar_install(dplyr)
}
}
