% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_psa.R
\name{tar_psa}
\alias{tar_psa}
\title{Peritoneal Surface Area Calculation}
\usage{
tar_psa(height, weight)
}
\arguments{
\item{height}{A numeric value representing the height in cm.}

\item{weight}{A numeric value representing the weight in kg.}
}
\value{
A numeric value representing 1% of the peritoneal surface area in cm^2.
}
\description{
This function calculates 1% of the peritoneal surface area in cm^2 based on
the given height and weight. It's used as a factor for further calculations.
}
\examples{
tar_psa(180, 80) # Expected output: 199.6421 cm^2
}
